Current changes:
-I implemented a logic into the MyTrainingScreen.kt searchBar. The user
is now allowed to make a textinput. When the user leaves
the current fragment, his preceding textinput get cleared and
the obtains the default phrase of "Suche". I am not
properly content with my logics in the searchbars. My created methods doesn't contain the usage of binding, because I didn't figure out, how to implement binding in this class. In the End there is still the impediment that I can't approach the ui views with binding, because I wasn't able to implement binding in this fragment.

Plans for continuing changes:
-Adjust a smaller padding in ui layouts beyond the bottom navigation top und the ui screen bottom.
-Do the first sketches for ui views which displays a listed array of weight training exercises.
-Do the logic for the other screens in this app.